pyproject.toml
[project]
name = "ecommerce-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain==0.3.27",
    "langchain-openai==0.3.33",
    "langchain-community==0.3.24",
    "faiss-cpu",
    "python-dotenv",
    "tiktoken"
]
